+++ src/main/java/volcengine/core/Constant.java
@@ -10,16 +10,11 @@ public final class Constant {
 
     public final static int MAX_IMPORT_ITEM_COUNT = 10000;
 
-    final static List<String> AIR_CN_HOSTS = Collections.singletonList("api.byteair.volces.com");
-
-    final static List<String> AIR_SG_HOSTS = Collections.singletonList("byteair-api-sg1.byteintlapi.com");
-
-    /**
+    /*
      * All requests will have a XXXResponse corresponding to them,
      * and all XXXResponses will contain a 'Status' field.
      * The status of this request can be determined by the value of `Status.Code`
-     * Detail error code info：https://docs.byteplus.com/docs/error-code
-     */
+    */
     // The request was executed successfully without any exception
     public final static int STATUS_CODE_SUCCESS = 0;
     // A Request with the same "Request-ID" was already received. This Request was rejected