+++ src/main/java/volcengine/core/Region.java
@@ -1,8 +1,0 @@
-package volcengine.core;
-
-public enum Region {
-    // byteair cn
-    AIR_CN,
-    // byteair sg
-    AIR_SG,
-}